extended tstruct for templating - tstruct is in new datastructure.go - updated fstab.ww to new syntax
