+++ overlays/wwinit/etc/fstab.ww
@@ -1,8 +1,20 @@
+# This file is autogenerated by warewulf 
+# Host:   {{.BuildHost}}
+# Time:   {{.BuildTime}}
+# Source: {{.BuildSource}}
 rootfs / tmpfs defaults 0 0
 devpts /dev/pts devpts gid=5,mode=620 0 0
 tmpfs /run/shm tmpfs defaults 0 0
 sysfs /sys sysfs defaults 0 0
 proc /proc proc defaults 0 0
-{{range .NFSMounts}}
-        {{.}}
-{{end}}
+{{- $IP:=.Ipaddr }}
+{{- if .Nfs.Enabled }}
+{{- range .Nfs.ExportsExtended }}
+{{- if .Mount }}
+{{ $IP }}:{{ .Path }} {{ .Path }} nfs {{ .MountOptions }} 0 0
+{{- end }}
+{{- end }}
+{{- end }}
+{{- range .Nfs.Exports }}
+{{ $IP }}:{{ . }} {{ . }} nfs defaults 0 0
+{{- end }}
